target/adi_v5_swd: make SWD connect DP lockup proof
An AP or a block/bus connected to it can be unpowered or
without a clock signal. In such situation DP locks up
on access to the AP or the non-functional block/bus.
In such state DP responses to IDR reads, CTRL_STAT reads
and ABORT writes by OK ack to comply with ADIv5.2 or v6
reference. All other accesses are refused with SWD WAIT.

Modify the (re)connect loops to respect SWD WAIT and keep
retrying until timeout and then issue DAPABORT to
revive the stalled DP.

Add direct write to DP SELECT register to swd_connect_single()
to detect stalled DP state and retry the connect loop.

Signed-off-by: Tomas Vanek <vanekt@fbl.cz>
Change-Id: I9d84792b6da08b967864c1ea4624a4fe8e9a006f