target/adi_v5_swd: switch to DORMANT to revive stalled DP
On several devices it was observed that DAPABORT bit
in ABORT DP register does not help to revive DP to functional
state. On the other hand, switching the DPv2 or v3
to DORMANT state reliably released the lock up and DP
can continue communicating to accessible APs.
Although it is not mentioned in ADI reference it
looks naturally that the DP logic gets reset after unpowering
or deep sleep in the DORMANT state.

Alternate issuing DAPABORT and switching to DORMANT state
after (re)connect timeout is over.

Signed-off-by: Tomas Vanek <vanekt@fbl.cz>
Change-Id: I9afdf971aa4b849ea48fdba40b748d0c7bd65130